+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-04-23 18:55:06
-// Commit: test(test): add unit tests for test business logic
+// Updated at: 2026-04-24 21:30:40
+// Commit: feat(test): add bi-directional event bus handler for test
 
 class testComponentState {
-  final String commitRef = '606';
-  final String lastMessage = 'test(test): add unit tests for test business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-23T18:55:06+05:00');
+  final String commitRef = '649';
+  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
+  final DateTime timestamp = DateTime.parse('2026-04-24T21:30:40+05:00');
 }